+++ src/services/custom_care_service/characteristics/care_filter.ts
@@ -1,13 +1,18 @@
-import { Characteristic } from "hap-nodejs";
+import type { HAP } from "homebridge";
 
-export class CareFilter extends Characteristic {
-  public static UUID = "00000102-0000-0000-0000-000000000000";
-  constructor() {
-    super("Care indicator filter", CareFilter.UUID, {
-      format: Characteristic.Formats.FLOAT,
-      unit: "%",
-      perms: [Characteristic.Perms.READ, Characteristic.Perms.NOTIFY],
-    });
-    this.value = this.getDefaultValue();
+export function createCareFilterClass(hap: HAP) {
+  class CareFilter extends hap.Characteristic {
+    public static UUID = "00000102-0000-0000-0000-000000000000";
+
+    constructor() {
+      super("Care indicator filter", CareFilter.UUID, {
+        format: hap.Characteristic.Formats.FLOAT,
+        unit: "%",
+        perms: [hap.Characteristic.Perms.READ, hap.Characteristic.Perms.NOTIFY],
+      });
+      this.value = this.getDefaultValue();
+    }
   }
+
+  return { CareFilter };
 }