+++ src/services/custom_care_service/characteristics/care_main_brush.ts
@@ -1,13 +1,16 @@
-import { Characteristic } from "hap-nodejs";
+import type { HAP } from "homebridge";
 
-export class CareMainBrush extends Characteristic {
-  public static UUID = "00000104-0000-0000-0000-000000000000";
-  constructor() {
-    super("Care indicator main brush", CareMainBrush.UUID, {
-      format: Characteristic.Formats.FLOAT,
-      unit: "%",
-      perms: [Characteristic.Perms.READ, Characteristic.Perms.NOTIFY],
-    });
-    this.value = this.getDefaultValue();
+export function createCareMainBrushClass(hap: HAP) {
+  class CareMainBrush extends hap.Characteristic {
+    public static UUID = "00000104-0000-0000-0000-000000000000";
+    constructor() {
+      super("Care indicator main brush", CareMainBrush.UUID, {
+        format: hap.Characteristic.Formats.FLOAT,
+        unit: "%",
+        perms: [hap.Characteristic.Perms.READ, hap.Characteristic.Perms.NOTIFY],
+      });
+      this.value = this.getDefaultValue();
+    }
   }
+  return { CareMainBrush };
 }