+++ src/services/custom_care_service/characteristics/care_sensors.ts
@@ -1,13 +1,17 @@
-import { Characteristic } from "hap-nodejs";
+import type { HAP } from "homebridge";
 
-export class CareSensors extends Characteristic {
-  public static UUID = "00000101-0000-0000-0000-000000000000";
-  constructor() {
-    super("Care indicator sensors", CareSensors.UUID, {
-      format: Characteristic.Formats.FLOAT,
-      unit: "%",
-      perms: [Characteristic.Perms.READ, Characteristic.Perms.NOTIFY],
-    });
-    this.value = this.getDefaultValue();
+export function createCareSensorsClass(hap: HAP) {
+  class CareSensors extends hap.Characteristic {
+    public static UUID = "00000101-0000-0000-0000-000000000000";
+
+    constructor() {
+      super("Care indicator sensors", CareSensors.UUID, {
+        format: hap.Characteristic.Formats.FLOAT,
+        unit: "%",
+        perms: [hap.Characteristic.Perms.READ, hap.Characteristic.Perms.NOTIFY],
+      });
+      this.value = this.getDefaultValue();
+    }
   }
+  return { CareSensors };
 }