+++ src/services/custom_care_service/characteristics/index.ts
@@ -1,4 +1,4 @@
-export { CareSensors } from "./care_sensors";
-export { CareFilter } from "./care_filter";
-export { CareMainBrush } from "./care_main_brush";
-export { CareSideBrush } from "./care_side_brush";
+export { createCareSensorsClass } from "./care_sensors";
+export { createCareFilterClass } from "./care_filter";
+export { createCareMainBrushClass } from "./care_main_brush";
+export { createCareSideBrushClass } from "./care_side_brush";