+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## Next
 
 - [Bug] Remove peer dependency on `hap-nodejs` as some folks had issues (#558)
+- [Bug] Accessories are randomly presented so the main service loses the primary status (#542)
 
 ## 0.26.5