+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## Next
 
+## 0.27.0
+
 - [Bug] Remove peer dependency on `hap-nodejs` as some folks had issues (#558)
 - [Bug] Accessories are randomly presented so the main service loses the primary status (#542)
 - [Bug] Reset cleaning status of Rooms and Zones after cleaning is completed (#546)